feat: using browserify  to bundle
- delete useless files
- add some build tools
