+++ build.js
@@ -1,0 +1,6 @@
+const fs = require('fs');
+const browserify = require('browserify');
+browserify('./lib/index.js')
+  .transform('babelify', { presets: ['@babel/preset-env'] })
+  .bundle()
+  .pipe(fs.createWriteStream('./dist/bundle.js'));
