--- lib/env.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const isNode = typeof process !== 'undefined' &&
-  process.versions != null &&
-  process.versions.node != null;
-
-module.exports = {
-  isNode
-};
