+++ lib/util.js
@@ -1,22 +1,15 @@
 'use strict';
 
 const qs = require('qs');
-const { isNode } = require('./env');
+const crypto = require('crypto');
+const randomBytes = crypto.randomBytes;
 
-let bufToHex;
-let hexToBuf;
-let randomBytes;
+const bufToHex = (buffer) => {
+  return buffer.toString('hex');
+};
 
-if (isNode) {
-  const nodeMethods = require('./node');
-  bufToHex = nodeMethods.bufToHex;
-  hexToBuf = nodeMethods.hexToBuf;
-  randomBytes = nodeMethods.randomBytes;
-} else {
-  const browserMethods = require('./browser');
-  bufToHex = browserMethods.bufToHex;
-  hexToBuf = browserMethods.hexToBuf;
-  randomBytes = browserMethods.randomBytes;
+const hexToBuf = (string) => {
+  return Buffer.from(string, 'hex');
 };
 
 const dumpBuf = (buffer, options = {}) => {
