+++ package.json
@@ -3,7 +3,7 @@
   "description": "prs utility",
   "version": "0.0.2",
   "homepage": "https://press.one",
-  "main": "./lib/utility.js",
+  "main": "./lib/index.js",
   "author": "PRESS.one",
   "repository": {
     "type": "git",
@@ -11,6 +11,7 @@
   },
   "scripts": {
     "test": "node test",
+    "build": "node build && uglifyjs ./dist/bundle.js > ./dist/bundle.min.js",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix"
   },
@@ -23,14 +24,19 @@
     "secp256k1": "^3.5.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.4.3",
+    "@babel/preset-env": "^7.4.3",
+    "babelify": "^10.0.0",
+    "browserify": "^16.2.3",
+    "chai": "^4.1.2",
     "eslint": "^5.16.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-node": "^8.0.1",
     "eslint-plugin-promise": "^4.1.1",
     "eslint-plugin-standard": "^4.0.0",
-    "chai": "^4.1.2",
     "mocha": "^5.0.2",
+    "uglify-js": "^3.5.3",
     "yargs": "^11.0.0"
   },
   "license": "MIT"
